requirements.txt
anyio==3.6.1
bcrypt==3.2.2
cffi==1.15.1
click==8.1.3
colorama==0.4.5
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.2.1
fastapi==0.79.0
h11==0.13.0
idna==3.3
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.2
pymongo==4.2.0
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.3.0
